requirements-dev.txt
pytest>=8.2
pytest-django>=4.9
pytest-cov>=5.0
model-bakery>=1.19
freezegun>=1.5
factory_boy>=3.3
respx>=0.21